Fixed the crash in R backend, but the backend can stop working instead.
When RSession::interrupt is called in the small interval between the
end of the calculation in the R server and the moment that Cantor recieves
the results, then the R server is killed.

+++ src/backends/R/rsession.cpp
@@ -73,7 +73,8 @@ void RSession::logout()
 void RSession::interrupt()
 {
     kDebug()<<"interrupt" << m_rProcess->pid();
-    kill(m_rProcess->pid(), 2);
+    if (m_rProcess->pid())
+	kill(m_rProcess->pid(), 2);
     m_expressionQueue.removeFirst();
     changeStatus(Cantor::Session::Done);
 }
